+++ src/commands/utility/reload.ts
@@ -5,7 +5,7 @@ import {
 	CommandHandler,
 	ListenerHandler,
 } from 'discord-akairo';
-import { messages } from '../../util/constants/messages';
+import { messages } from '../../util/constant/messages';
 
 export default class ReloadCommand extends Command {
 	constructor() {
